pyproject.toml
[tool.poetry]
name = "hdfc-analytics"
version = "0.3.0"
description = "HDFC statement analysis"
authors = ["Vipul Sharma <vipul.sharma20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.3"
toml = "^0.10.2"
docopt = "^0.6.2"
plotly = "^5.18.0"
litellm = "^1.61.6"
hdfc-cc-parser-rs = "^0.0.8"

[tool.poetry.scripts]
hdfc-analytics = "hdfc_analytics.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
